+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## unreleased
 
+* Prevent a user from updating their currently active subscription to the same billing plan, unless the
+  subscription is on a grace period (ie. will cancel at period end).
+
 ## 8.x-2.0
 
 * Initial stable release.